fix: revert "fix: split linear algebra into one file per class"
This reverts commit 93dbc3be84559cd121b8e468e6a8cfebcfd84ba4.
