+++ src/CMakeLists.txt
@@ -39,8 +39,7 @@ set(SOURCES
     io/buffer_exporter.cpp
     ipc/message_exchange.cpp
     ipc/raw_data_decode.cpp
-    math/mat4.cpp
-    math/vec4.cpp
+    math/linear_algebra.cpp
     ui/decorated_line_edit.cpp
     ui/gl_canvas.cpp
     ui/gl_text_renderer.cpp
