+++ src/math/linear_algebra.cpp
@@ -23,13 +23,101 @@
  * IN THE SOFTWARE.
  */
 
-#include "mat4.h"
+#include "linear_algebra.h"
 
-#include <initializer_list>
 #include <iostream>
+
 
 namespace oid
 {
+
+
+vec4::vec4(const float x, const float y, const float z, const float w)
+    : vec_{x, y, z, w}
+{
+}
+
+
+vec4& vec4::operator+=(const vec4& b)
+{
+    for (int i = 0; i < 4; ++i) {
+        vec_[i] += b.vec_[i];
+    }
+    return *this;
+}
+
+
+void vec4::print() const
+{
+    std::cout << vec_.transpose() << std::endl;
+}
+
+
+float* vec4::data()
+{
+    return vec_.data();
+}
+
+
+float& vec4::x()
+{
+    return vec_[0];
+}
+
+
+float& vec4::y()
+{
+    return vec_[1];
+}
+
+
+float& vec4::z()
+{
+    return vec_[2];
+}
+
+
+float& vec4::w()
+{
+    return vec_[3];
+}
+
+
+const float& vec4::x() const
+{
+    return vec_[0];
+}
+
+
+const float& vec4::y() const
+{
+    return vec_[1];
+}
+
+
+const float& vec4::z() const
+{
+    return vec_[2];
+}
+
+
+const float& vec4::w() const
+{
+    return vec_[3];
+}
+
+
+vec4 vec4::zero()
+{
+    return {0.0f, 0.0f, 0.0f, 0.0f};
+}
+
+
+vec4 operator-(const vec4& vector)
+{
+    return {-vector.x(), -vector.y(), -vector.z(), -vector.w()};
+}
+
 
 void mat4::set_identity()
 {
